fixup! Fixup mock dtypes

+++ tests/test_hstrat/test_juxtaposition/test_impl_specimen/test_calc_rank_of_first_retained_disparity_between_specimen_naive.py
@@ -64,8 +64,12 @@ def specimens_mock_complex():
     second = MagicMock()
     first.GetStratumDifferentiaBitWidth.return_value = 8
     second.GetStratumDifferentiaBitWidth.return_value = 8
-    first.GetDifferentiaVals.return_value = np.array([5, 6, 7, 8, 9])
-    second.GetDifferentiaVals.return_value = np.array([5, 42, 9, 202])
+    first.GetDifferentiaVals.return_value = np.array(
+        [5, 6, 7, 8, 9], dtype=np.uint32
+    )
+    second.GetDifferentiaVals.return_value = np.array(
+        [5, 42, 9, 202], dtype=np.uint32
+    )
     first.GetNumStrataRetained.return_value = 5
     second.GetNumStrataRetained.return_value = 4
     first.GetNumStrataDeposited.return_value = 5
